requirements.txt
gradio>=4.0.0
chromadb>=0.4.0
openai>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
